Implement QueryMonitor callbacks for Presto on Spark
