+++ presto-main/src/main/java/com/facebook/presto/execution/StageExecutionStateMachine.java
@@ -18,20 +18,15 @@ import com.facebook.airlift.stats.Distribution;
 import com.facebook.presto.execution.StateMachine.StateChangeListener;
 import com.facebook.presto.execution.scheduler.SplitSchedulerStats;
 import com.facebook.presto.operator.BlockedReason;
-import com.facebook.presto.operator.OperatorStats;
-import com.facebook.presto.operator.PipelineStats;
 import com.facebook.presto.operator.TaskStats;
-import com.facebook.presto.spi.eventlistener.StageGcStatistics;
 import com.facebook.presto.util.Failures;
 import com.google.common.collect.ImmutableList;
 import org.joda.time.DateTime;
 
 import javax.annotation.concurrent.ThreadSafe;
 
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.OptionalDouble;
 import java.util.Set;
@@ -56,14 +51,11 @@ import static com.google.common.base.MoreObjects.toStringHelper;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkState;
 import static io.airlift.units.DataSize.succinctBytes;
-import static io.airlift.units.Duration.succinctDuration;
 import static io.airlift.units.Duration.succinctNanos;
 import static java.lang.Math.max;
 import static java.lang.Math.min;
-import static java.lang.Math.toIntExact;
 import static java.util.Objects.requireNonNull;
 import static java.util.concurrent.TimeUnit.NANOSECONDS;
-import static java.util.concurrent.TimeUnit.SECONDS;
 
 @ThreadSafe
 public class StageExecutionStateMachine
@@ -334,176 +326,20 @@ public class StageExecutionStateMachine
         StageExecutionState state = this.state.get();
 
         List<TaskInfo> taskInfos = ImmutableList.copyOf(taskInfosSupplier.get());
-
-        int totalTasks = taskInfos.size();
-        int runningTasks = 0;
-        int completedTasks = 0;
-
-        int totalDrivers = 0;
-        int queuedDrivers = 0;
-        int runningDrivers = 0;
-        int blockedDrivers = 0;
-        int completedDrivers = 0;
-
-        long cumulativeUserMemory = 0;
-        long userMemoryReservation = 0;
-        long totalMemoryReservation = 0;
-        long peakUserMemoryReservation = peakUserMemory.get();
-
-        long totalScheduledTime = 0;
-        long totalCpuTime = 0;
-        long retriedCpuTime = 0;
-        long totalBlockedTime = 0;
-
-        long totalAllocation = 0;
-
-        long rawInputDataSize = 0;
-        long rawInputPositions = 0;
-
-        long processedInputDataSize = 0;
-        long processedInputPositions = 0;
-
-        long bufferedDataSize = 0;
-        long outputDataSize = 0;
-        long outputPositions = 0;
-
-        long physicalWrittenDataSize = 0;
-
-        int fullGcCount = 0;
-        int fullGcTaskCount = 0;
-        int minFullGcSec = 0;
-        int maxFullGcSec = 0;
-        int totalFullGcSec = 0;
-
-        boolean fullyBlocked = true;
-        Set<BlockedReason> blockedReasons = new HashSet<>();
-
-        Map<String, OperatorStats> operatorToStats = new HashMap<>();
-        for (TaskInfo taskInfo : taskInfos) {
-            TaskState taskState = taskInfo.getTaskStatus().getState();
-            if (taskState.isDone()) {
-                completedTasks++;
-            }
-            else {
-                runningTasks++;
-            }
-
-            TaskStats taskStats = taskInfo.getStats();
-
-            totalDrivers += taskStats.getTotalDrivers();
-            queuedDrivers += taskStats.getQueuedDrivers();
-            runningDrivers += taskStats.getRunningDrivers();
-            blockedDrivers += taskStats.getBlockedDrivers();
-            completedDrivers += taskStats.getCompletedDrivers();
-
-            cumulativeUserMemory += taskStats.getCumulativeUserMemory();
-
-            long taskUserMemory = taskStats.getUserMemoryReservation().toBytes();
-            long taskSystemMemory = taskStats.getSystemMemoryReservation().toBytes();
-            userMemoryReservation += taskUserMemory;
-            totalMemoryReservation += taskUserMemory + taskSystemMemory;
-
-            totalScheduledTime += taskStats.getTotalScheduledTime().roundTo(NANOSECONDS);
-            totalCpuTime += taskStats.getTotalCpuTime().roundTo(NANOSECONDS);
-            if (state == FINISHED && taskInfo.getTaskStatus().getState() == TaskState.FAILED) {
-                retriedCpuTime += taskStats.getTotalCpuTime().roundTo(NANOSECONDS);
-            }
-            totalBlockedTime += taskStats.getTotalBlockedTime().roundTo(NANOSECONDS);
-            if (!taskState.isDone()) {
-                fullyBlocked &= taskStats.isFullyBlocked();
-                blockedReasons.addAll(taskStats.getBlockedReasons());
-            }
-
-            totalAllocation += taskStats.getTotalAllocation().toBytes();
-
-            rawInputDataSize += taskStats.getRawInputDataSize().toBytes();
-            rawInputPositions += taskStats.getRawInputPositions();
-
-            processedInputDataSize += taskStats.getProcessedInputDataSize().toBytes();
-            processedInputPositions += taskStats.getProcessedInputPositions();
-
-            bufferedDataSize += taskInfo.getOutputBuffers().getTotalBufferedBytes();
-            outputDataSize += taskStats.getOutputDataSize().toBytes();
-            outputPositions += taskStats.getOutputPositions();
-
-            physicalWrittenDataSize += taskStats.getPhysicalWrittenDataSize().toBytes();
-
-            fullGcCount += taskStats.getFullGcCount();
-            fullGcTaskCount += taskStats.getFullGcCount() > 0 ? 1 : 0;
-
-            int gcSec = toIntExact(taskStats.getFullGcTime().roundTo(SECONDS));
-            totalFullGcSec += gcSec;
-            minFullGcSec = min(minFullGcSec, gcSec);
-            maxFullGcSec = max(maxFullGcSec, gcSec);
-
-            for (PipelineStats pipeline : taskStats.getPipelines()) {
-                for (OperatorStats operatorStats : pipeline.getOperatorSummaries()) {
-                    String id = pipeline.getPipelineId() + "." + operatorStats.getOperatorId();
-                    operatorToStats.compute(id, (k, v) -> v == null ? operatorStats : v.add(operatorStats));
-                }
-            }
-        }
-
-        StageExecutionStats stageExecutionStats = new StageExecutionStats(
-                schedulingComplete.get(),
-                getSplitDistribution.snapshot(),
-
-                totalTasks,
-                runningTasks,
-                completedTasks,
-
-                totalLifespans,
-                finishedLifespans,
-
-                totalDrivers,
-                queuedDrivers,
-                runningDrivers,
-                blockedDrivers,
-                completedDrivers,
-
-                cumulativeUserMemory,
-                succinctBytes(userMemoryReservation),
-                succinctBytes(totalMemoryReservation),
-                succinctBytes(peakUserMemoryReservation),
-                succinctDuration(totalScheduledTime, NANOSECONDS),
-                succinctDuration(totalCpuTime, NANOSECONDS),
-                succinctDuration(retriedCpuTime, NANOSECONDS),
-                succinctDuration(totalBlockedTime, NANOSECONDS),
-                fullyBlocked && runningTasks > 0,
-                blockedReasons,
-
-                succinctBytes(totalAllocation),
-
-                succinctBytes(rawInputDataSize),
-                rawInputPositions,
-                succinctBytes(processedInputDataSize),
-                processedInputPositions,
-                succinctBytes(bufferedDataSize),
-                succinctBytes(outputDataSize),
-                outputPositions,
-                succinctBytes(physicalWrittenDataSize),
-
-                new StageGcStatistics(
-                        stageExecutionId.getStageId().getId(),
-                        stageExecutionId.getId(),
-                        totalTasks,
-                        fullGcTaskCount,
-                        minFullGcSec,
-                        maxFullGcSec,
-                        totalFullGcSec,
-                        (int) (1.0 * totalFullGcSec / fullGcCount)),
-
-                ImmutableList.copyOf(operatorToStats.values()));
-
         Optional<ExecutionFailureInfo> failureInfo = Optional.empty();
         if (state == FAILED) {
             failureInfo = Optional.of(failureCause.get());
         }
-        return new StageExecutionInfo(
+        return StageExecutionInfo.create(
+                stageExecutionId,
                 state,
-                stageExecutionStats,
+                failureInfo,
                 taskInfos,
-                failureInfo);
+                schedulingComplete.get(),
+                getSplitDistribution.snapshot(),
+                succinctBytes(peakUserMemory.get()),
+                finishedLifespans,
+                totalLifespans);
     }
 
     public void recordGetSplitTime(long startNanos)
