+++ presto-spark-base/src/main/java/com/facebook/presto/spark/planner/PrestoSparkQueryPlanner.java
@@ -16,24 +16,35 @@ package com.facebook.presto.spark.planner;
 import com.facebook.presto.Session;
 import com.facebook.presto.cost.CostCalculator;
 import com.facebook.presto.cost.StatsCalculator;
+import com.facebook.presto.execution.Input;
+import com.facebook.presto.execution.Output;
 import com.facebook.presto.execution.QueryPreparer.PreparedQuery;
 import com.facebook.presto.execution.warnings.WarningCollector;
 import com.facebook.presto.metadata.Metadata;
 import com.facebook.presto.security.AccessControl;
 import com.facebook.presto.spi.plan.PlanNodeIdAllocator;
+import com.facebook.presto.spi.resourceGroups.QueryType;
 import com.facebook.presto.sql.analyzer.Analysis;
 import com.facebook.presto.sql.analyzer.Analyzer;
 import com.facebook.presto.sql.analyzer.QueryExplainer;
 import com.facebook.presto.sql.parser.SqlParser;
+import com.facebook.presto.sql.planner.InputExtractor;
 import com.facebook.presto.sql.planner.LogicalPlanner;
+import com.facebook.presto.sql.planner.OutputExtractor;
 import com.facebook.presto.sql.planner.Plan;
 import com.facebook.presto.sql.planner.PlanOptimizers;
+import com.facebook.presto.sql.planner.plan.OutputNode;
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableSet;
 
 import javax.inject.Inject;
 
+import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 import static com.facebook.presto.sql.planner.LogicalPlanner.Stage.OPTIMIZED_AND_VALIDATED;
+import static com.facebook.presto.util.StatementUtils.getQueryType;
 import static java.util.Objects.requireNonNull;
 
 public class PrestoSparkQueryPlanner
@@ -65,7 +76,7 @@ public class PrestoSparkQueryPlanner
         this.accessControl = requireNonNull(accessControl, "accessControl is null");
     }
 
-    public PlanAndUpdateType createQueryPlan(Session session, PreparedQuery preparedQuery, WarningCollector warningCollector)
+    public PlanAndMore createQueryPlan(Session session, PreparedQuery preparedQuery, WarningCollector warningCollector)
     {
         PlanNodeIdAllocator idAllocator = new PlanNodeIdAllocator();
 
@@ -90,20 +101,43 @@ public class PrestoSparkQueryPlanner
                 warningCollector);
 
         Analysis analysis = analyzer.analyze(preparedQuery.getStatement());
-        return new PlanAndUpdateType(
-                logicalPlanner.plan(analysis, OPTIMIZED_AND_VALIDATED),
-                Optional.ofNullable(analysis.getUpdateType()));
+        Plan plan = logicalPlanner.plan(analysis, OPTIMIZED_AND_VALIDATED);
+        List<Input> inputs = new InputExtractor(metadata, session).extractInputs(plan.getRoot());
+        Optional<Output> output = new OutputExtractor().extractOutput(plan.getRoot());
+        Optional<QueryType> queryType = getQueryType(preparedQuery.getStatement().getClass());
+        List<String> columnNames = ((OutputNode) plan.getRoot()).getColumnNames();
+        return new PlanAndMore(
+                plan,
+                Optional.ofNullable(analysis.getUpdateType()),
+                columnNames,
+                ImmutableSet.copyOf(inputs),
+                output,
+                queryType);
     }
 
-    public static class PlanAndUpdateType
+    public static class PlanAndMore
     {
         private final Plan plan;
         private final Optional<String> updateType;
+        private final List<String> fieldNames;
+        private final Set<Input> inputs;
+        private final Optional<Output> output;
+        private final Optional<QueryType> queryType;
 
-        public PlanAndUpdateType(Plan plan, Optional<String> updateType)
+        public PlanAndMore(
+                Plan plan,
+                Optional<String> updateType,
+                List<String> fieldNames,
+                Set<Input> inputs,
+                Optional<Output> output,
+                Optional<QueryType> queryType)
         {
             this.plan = requireNonNull(plan, "plan is null");
             this.updateType = requireNonNull(updateType, "updateType is null");
+            this.fieldNames = ImmutableList.copyOf(requireNonNull(fieldNames, "fieldNames is null"));
+            this.inputs = ImmutableSet.copyOf(requireNonNull(inputs, "inputs is null"));
+            this.output = requireNonNull(output, "output is null");
+            this.queryType = requireNonNull(queryType, "queryType is null");
         }
 
         public Plan getPlan()
@@ -114,6 +148,26 @@ public class PrestoSparkQueryPlanner
         public Optional<String> getUpdateType()
         {
             return updateType;
+        }
+
+        public List<String> getFieldNames()
+        {
+            return fieldNames;
+        }
+
+        public Set<Input> getInputs()
+        {
+            return inputs;
+        }
+
+        public Optional<Output> getOutput()
+        {
+            return output;
+        }
+
+        public Optional<QueryType> getQueryType()
+        {
+            return queryType;
         }
     }
 }
